requirements.txt
asgiref==3.5.1
black==22.3.0
click==8.1.3
Django==4.0.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
whitenoise==6.0.0
